Create Username Changes
A company has released a new internal system, and each employee has been assigned a username. Employees are allowed to change their usernames but only in a limited way. More specifically, they can choose letters at two different positions and swap them. For example, the username "bigfish" can be changed to "gibfish" (swapping 'b' and 'g') or "bighisf" (swapping 'f' and 'h'). The manager would like to know which employees can update their usernames so that the new username is smaller in alphabetical order than the original username.

 

For each username given, return either "YES" or "NO" based on whether that username can be changed (with one swap) to a new one that is smaller in alphabetical order.

 

Note: For two different strings A and B of the same length, A is smaller than B in alphabetical order when on the first position where A and B differ, A has a smaller letter in alphabetical order than B has.

 

For example, let's say usernames = ["bee", "superhero", "ace"]. For the first username, "bee", it is not possible to make one swap to change it to a smaller one in alphabetical order, so the answer is "NO". For the second username, "superhero", it is possible get a new username that is smaller in alphabetical order (for example, by swapping letters 's' and 'h' to get "hupersero"), so the answer is "YES". Finally, for the last username "ace", it is not possible to make one swap to change it to a smaller one in alphabetical order, so the answer is "NO". Therefore you would return the array of strings ["NO", "YES", "NO"].

 
Function Description

Complete the function possibleChanges in the editor below.

 

possibleChanges has the following parameter(s):

    string usernames[n]:  an array of strings denoting the usernames of the employees

Returns:

    string[n]: an array of strings containing either "YES" or "NO" based on whether the ith username can be changed with one swap to a new one that is smaller in alphabetical order

 

Constraints

1 ≤ n  ≤ 105

The sum of lengths of all usernames does not exceed 106.

usernames[i] consists of only lowercase English letters.